Fix crash in audio source
Pulse Audio source is currently the only source that has its own thread
to poll for changes. Embedded Lua is not thread safe so at some cases
publishing audio source changes on that thread caused the program to
crash. This refactors to make sure that publishing of source states to
Lua always happens on main thread.